+++ lib/formatters/formatter_extension_methods.dart
@@ -74,6 +74,14 @@ extension NumericInputFormatting on num {
 }
 
 extension StringInputFormatting on String {
+  bool get isFiatCurrency {
+    return fu.isFiatCurrency(this);
+  }
+
+  bool get isCryptoCurrency {
+    return fu.isCryptoCurrency(this);
+  }
+
   /// [thousandSeparator] specifies what symbol will be used to separate
   /// each block of 3 digits, e.g. [ThousandSeparator.Comma] will format
   /// a million as 1,000,000