+++ lib/formatters/formatter_utils.dart
@@ -24,6 +24,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 import 'package:flutter/foundation.dart';
+import 'package:flutter_multi_formatter/formatters/all_fiat_currencies.dart';
 
 import 'money_input_enums.dart';
 
@@ -455,6 +456,27 @@ String _getRoundedValue(
     return prepared;
   }
   return result.toInt().toString();
+}
+
+/// Checks if currency is fiat
+bool isFiatCurrency(String currencyId) {
+  if (currencyId.length != 3) {
+    return false;
+  }
+  return allFiatCurrencies.contains(
+    currencyId.toUpperCase(),
+  );
+}
+
+/// Basically it doesn't really check if the currencyId is
+/// a crypto currency. It just checks if it's not fiat.
+/// I decided not to collect all possible crypto currecies as
+/// there's an endless amount of them
+bool isCryptoCurrency(String currencyId) {
+  if (currencyId.length < 3 || currencyId.length > 4) {
+    return false;
+  }
+  return !isFiatCurrency(currencyId);
 }
 
 /// simply adds a period to an existing fractional part